+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Download button label translation for Resource paragraph.
 - Added submodule which provides default configuration for LMS search functionality.
 - Fixed course navigation overlapping by tables on mobiles.
+- Added search keywords highlighting i lesson content.
 
 ## [2.5.6]
 - Disabled cache when the progress is enabled.